// services/Trader-Trainer/src/core/neural_network/type.ts
export type ActivationType =
  | "sigmoid"
  | "tanh"
  | "ReLu"
  | "leakyReLu"
  | "GELU"
  | "softmax"
  | "ELU"
  | "mish";